refactor(extensions): split scheduler module
