+++ .changeset/split-scheduler-into-focused-modules.md
@@ -1,0 +1,5 @@
+---
+default: patch
+---
+
+refactor the scheduler extension into shared constants/types, parsing helpers, and registration modules without changing user-facing behavior.
