+++ packages/extensions/extensions/scheduler.ts
@@ -8,451 +8,80 @@
  *
  * Based on pi-scheduler by @manojlds (MIT).
  *
- * Tasks run only while pi is active and idle. Recurring tasks auto-expire
- * after 3 days. State is persisted under `~/.pi/agent/scheduler/.../scheduler.json`
- * using a path that mirrors the current workspace path.
+ * Tasks run only while pi is active and idle. State is persisted under
+ * `~/.pi/agent/scheduler/.../scheduler.json` using a path that mirrors the
+ * current workspace path.
  */
 
 import * as fs from "node:fs";
 import * as path from "node:path";
-import { type ExtensionAPI, type ExtensionContext, getAgentDir } from "@mariozechner/pi-coding-agent";
-import { Type } from "@sinclair/typebox";
-import { Cron } from "croner";
+import type { ExtensionAPI, ExtensionContext } from "@mariozechner/pi-coding-agent";
+import {
+	computeNextCronRunAt,
+	formatDurationShort,
+	normalizeCronExpression,
+	normalizeDuration,
+	parseDuration,
+} from "./scheduler-parsing.js";
+import { registerCommands, registerEvents, registerTools } from "./scheduler-registration.js";
+import {
+	DEFAULT_LOOP_INTERVAL,
+	DISPATCH_RATE_LIMIT_WINDOW_MS,
+	FIFTEEN_MINUTES,
+	getLegacySchedulerStoragePath,
+	getSchedulerStoragePath,
+	getSchedulerStorageRoot,
+	MAX_DISPATCHES_PER_WINDOW,
+	MAX_TASKS,
+	MIN_RECURRING_INTERVAL,
+	ONE_MINUTE,
+	type ParseResult,
+	type RecurringSpec,
+	type ReminderParseResult,
+	type SchedulePromptAddPlan,
+	type ScheduleTask,
+	type TaskKind,
+	type TaskStatus,
+	THREE_DAYS,
+} from "./scheduler-shared.js";
 
-// ── Constants ───────────────────────────────────────────────────────────────
-
-export const MAX_TASKS = 50;
-export const ONE_MINUTE = 60_000;
-export const FIFTEEN_MINUTES = 15 * ONE_MINUTE;
-export const THREE_DAYS = 3 * 24 * 60 * ONE_MINUTE;
-export const DEFAULT_LOOP_INTERVAL = 10 * ONE_MINUTE;
-export const MIN_RECURRING_INTERVAL = ONE_MINUTE;
-export const DISPATCH_RATE_LIMIT_WINDOW_MS = ONE_MINUTE;
-export const MAX_DISPATCHES_PER_WINDOW = 6;
-
-// ── Types ───────────────────────────────────────────────────────────────────
-
-export type TaskKind = "recurring" | "once";
-export type TaskStatus = "pending" | "success" | "error";
-
-export interface ScheduleTask {
-	id: string;
-	prompt: string;
-	kind: TaskKind;
-	enabled: boolean;
-	createdAt: number;
-	nextRunAt: number;
-	intervalMs?: number;
-	cronExpression?: string;
-	expiresAt?: number;
-	jitterMs: number;
-	lastRunAt?: number;
-	lastStatus?: TaskStatus;
-	runCount: number;
-	pending: boolean;
-}
-
-export type RecurringSpec =
-	| { mode: "interval"; durationMs: number; note?: string }
-	| { mode: "cron"; cronExpression: string; note?: string };
-
-export interface ParseResult {
-	prompt: string;
-	recurring: RecurringSpec;
-}
-
-export interface ReminderParseResult {
-	prompt: string;
-	durationMs: number;
-	note?: string;
-}
-
-export type SchedulePromptAddPlan =
-	| { kind: "once"; durationMs: number; note?: string }
-	| { kind: "recurring"; mode: "interval"; durationMs: number; note?: string }
-	| { kind: "recurring"; mode: "cron"; cronExpression: string; note?: string };
+export {
+	computeCronCadenceMs,
+	computeNextCronRunAt,
+	formatDurationShort,
+	normalizeCronExpression,
+	normalizeDuration,
+	parseDuration,
+	parseLoopScheduleArgs,
+	parseRemindScheduleArgs,
+	validateSchedulePromptAddInput,
+} from "./scheduler-parsing.js";
+export {
+	DEFAULT_LOOP_INTERVAL,
+	DISPATCH_RATE_LIMIT_WINDOW_MS,
+	FIFTEEN_MINUTES,
+	getLegacySchedulerStoragePath,
+	getSchedulerStoragePath,
+	getSchedulerStorageRoot,
+	MAX_DISPATCHES_PER_WINDOW,
+	MAX_TASKS,
+	MIN_RECURRING_INTERVAL,
+	ONE_MINUTE,
+	THREE_DAYS,
+};
+export type {
+	ParseResult,
+	RecurringSpec,
+	ReminderParseResult,
+	SchedulePromptAddPlan,
+	ScheduleTask,
+	TaskKind,
+	TaskStatus,
+};
 
 interface SchedulerStore {
 	version: 1;
 	tasks: ScheduleTask[];
-}
-
-function getSchedulerStorageRoot(): string {
-	return path.join(getAgentDir(), "scheduler");
-}
-
-export function getSchedulerStoragePath(cwd: string): string {
-	const resolved = path.resolve(cwd);
-	const parsed = path.parse(resolved);
-	const relativeSegments = resolved.slice(parsed.root.length).split(path.sep).filter(Boolean);
-	const rootSegment = parsed.root
-		? parsed.root
-				.replaceAll(/[^a-zA-Z0-9]+/g, "-")
-				.replaceAll(/^-+|-+$/g, "")
-				.toLowerCase() || "root"
-		: "root";
-	return path.join(getSchedulerStorageRoot(), rootSegment, ...relativeSegments, "scheduler.json");
-}
-
-function getLegacySchedulerStoragePath(cwd: string): string {
-	return path.join(cwd, ".pi", "scheduler.json");
-}
-
-// ── Scheduling helpers ──────────────────────────────────────────────────────
-
-export function normalizeCronExpression(rawInput: string): { expression: string; note?: string } | undefined {
-	const input = rawInput.trim();
-	if (!input) {
-		return undefined;
-	}
-
-	const fields = input.split(/\s+/).filter(Boolean);
-	if (fields.length !== 5 && fields.length !== 6) {
-		return undefined;
-	}
-
-	const expression = fields.length === 5 ? `0 ${fields.join(" ")}` : fields.join(" ");
-	try {
-		// biome-ignore lint/suspicious/noEmptyBlockStatements: Cron requires a callback
-		const cron = new Cron(expression, () => {});
-		cron.stop();
-
-		const cadenceMs = computeCronCadenceMs(expression);
-		if (cadenceMs !== undefined && cadenceMs < MIN_RECURRING_INTERVAL) {
-			return undefined;
-		}
-
-		return {
-			expression,
-			note: fields.length === 5 ? "Interpreted as 5-field cron and normalized by prepending seconds=0." : undefined,
-		};
-	} catch {
-		return undefined;
-	}
-}
-
-export function computeNextCronRunAt(expression: string, fromTs = Date.now()): number | undefined {
-	try {
-		// biome-ignore lint/suspicious/noEmptyBlockStatements: Cron requires a callback
-		const cron = new Cron(expression, () => {});
-		const next = cron.nextRun(new Date(fromTs));
-		cron.stop();
-		return next?.getTime();
-	} catch {
-		return undefined;
-	}
-}
-
-export function computeCronCadenceMs(expression: string, fromTs = Date.now()): number | undefined {
-	try {
-		// biome-ignore lint/suspicious/noEmptyBlockStatements: Cron requires a callback
-		const cron = new Cron(expression, () => {});
-		const firstRun = cron.nextRun(new Date(fromTs));
-		if (!firstRun) {
-			cron.stop();
-			return undefined;
-		}
-		const secondRun = cron.nextRun(new Date(firstRun.getTime() + 1));
-		cron.stop();
-		if (!secondRun) {
-			return undefined;
-		}
-		return secondRun.getTime() - firstRun.getTime();
-	} catch {
-		return undefined;
-	}
-}
-
-export function formatDurationShort(ms: number): string {
-	if (ms % (24 * 60 * ONE_MINUTE) === 0) {
-		return `${ms / (24 * 60 * ONE_MINUTE)}d`;
-	}
-	if (ms % (60 * ONE_MINUTE) === 0) {
-		return `${ms / (60 * ONE_MINUTE)}h`;
-	}
-	return `${ms / ONE_MINUTE}m`;
-}
-
-export function normalizeDuration(durationMs: number): { durationMs: number; note?: string } {
-	if (durationMs <= 0) {
-		return { durationMs: ONE_MINUTE, note: "Rounded up to 1m (minimum interval)." };
-	}
-
-	const rounded = Math.ceil(durationMs / ONE_MINUTE) * ONE_MINUTE;
-	if (rounded !== durationMs) {
-		return {
-			durationMs: rounded,
-			note: `Rounded to ${formatDurationShort(rounded)} (minute granularity).`,
-		};
-	}
-	return { durationMs };
-}
-
-export function parseDuration(text: string): number | undefined {
-	const raw = text.trim().toLowerCase();
-	if (!raw) {
-		return undefined;
-	}
-
-	let match = raw.match(/^(\d+)\s*([smhd])$/i);
-	if (match) {
-		const n = Number.parseInt(match[1], 10);
-		const unit = match[2].toLowerCase();
-		if (unit === "s") {
-			return n * 1000;
-		}
-		if (unit === "m") {
-			return n * ONE_MINUTE;
-		}
-		if (unit === "h") {
-			return n * 60 * ONE_MINUTE;
-		}
-		if (unit === "d") {
-			return n * 24 * 60 * ONE_MINUTE;
-		}
-	}
-
-	match = raw.match(/^(\d+)\s*(seconds?|secs?|minutes?|mins?|hours?|hrs?|days?)$/i);
-	if (!match) {
-		return undefined;
-	}
-	const n = Number.parseInt(match[1], 10);
-	const unit = match[2].toLowerCase();
-	if (unit.startsWith("sec")) {
-		return n * 1000;
-	}
-	if (unit.startsWith("min")) {
-		return n * ONE_MINUTE;
-	}
-	if (unit.startsWith("hour") || unit.startsWith("hr")) {
-		return n * 60 * ONE_MINUTE;
-	}
-	if (unit.startsWith("day")) {
-		return n * 24 * 60 * ONE_MINUTE;
-	}
-	return undefined;
-}
-
-function extractLeadingDuration(input: string): { durationMs: number; prompt: string } | undefined {
-	const tokens = input.trim().split(/\s+/);
-	if (tokens.length < 2) {
-		return undefined;
-	}
-
-	const maxPrefix = Math.min(3, tokens.length - 1);
-	for (let i = 1; i <= maxPrefix; i++) {
-		const durationCandidate = tokens.slice(0, i).join(" ");
-		const durationMs = parseDuration(durationCandidate);
-		if (!durationMs) {
-			continue;
-		}
-		const prompt = tokens.slice(i).join(" ").trim();
-		if (!prompt) {
-			continue;
-		}
-		return { durationMs, prompt };
-	}
-
-	return undefined;
-}
-
-function extractLeadingCron(input: string): { cronExpression: string; prompt: string; note?: string } | undefined {
-	const trimmed = input.trim();
-	if (!trimmed.toLowerCase().startsWith("cron ")) {
-		return undefined;
-	}
-
-	const rest = trimmed.slice(5).trim();
-	if (!rest) {
-		return undefined;
-	}
-
-	const quotedMatch = rest.match(/^(["'])(.+?)\1\s+(.+)$/);
-	if (quotedMatch) {
-		const normalized = normalizeCronExpression(quotedMatch[2]);
-		const prompt = quotedMatch[3].trim();
-		if (!(normalized && prompt)) {
-			return undefined;
-		}
-		return { cronExpression: normalized.expression, prompt, note: normalized.note };
-	}
-
-	const tokens = rest.split(/\s+/);
-	for (const fieldCount of [6, 5]) {
-		if (tokens.length <= fieldCount) {
-			continue;
-		}
-
-		if (fieldCount === 5 && tokens.length >= 6) {
-			const sixthToken = tokens[5];
-			const sixthTokenLooksLikeCronField =
-				/^[\d*/?,#LWH-]+$/i.test(sixthToken) || /^[A-Z]{1,3}(?:,[A-Z]{1,3})*$/i.test(sixthToken);
-			if (sixthTokenLooksLikeCronField) {
-				continue;
-			}
-		}
-
-		const expressionCandidate = tokens.slice(0, fieldCount).join(" ");
-		const normalized = normalizeCronExpression(expressionCandidate);
-		if (!normalized) {
-			continue;
-		}
-		const prompt = tokens.slice(fieldCount).join(" ").trim();
-		if (!prompt) {
-			continue;
-		}
-		return { cronExpression: normalized.expression, prompt, note: normalized.note };
-	}
-
-	return undefined;
-}
-
-export function parseLoopScheduleArgs(args: string): ParseResult | undefined {
-	const input = args.trim();
-	if (!input) {
-		return undefined;
-	}
-
-	const explicitlyCron = input.toLowerCase().startsWith("cron ");
-	const leadingCron = extractLeadingCron(input);
-	if (leadingCron) {
-		return {
-			prompt: leadingCron.prompt,
-			recurring: {
-				mode: "cron",
-				cronExpression: leadingCron.cronExpression,
-				note: leadingCron.note,
-			},
-		};
-	}
-	if (explicitlyCron) {
-		return undefined;
-	}
-
-	const leading = extractLeadingDuration(input);
-	if (leading) {
-		const normalized = normalizeDuration(leading.durationMs);
-		return {
-			prompt: leading.prompt,
-			recurring: {
-				mode: "interval",
-				durationMs: normalized.durationMs,
-				note: normalized.note,
-			},
-		};
-	}
-
-	const trailingEvery = input.match(/^(.*)\s+every\s+(.+)$/i);
-	if (trailingEvery) {
-		const prompt = trailingEvery[1].trim();
-		const parsed = parseDuration(trailingEvery[2]);
-		if (prompt && parsed) {
-			const normalized = normalizeDuration(parsed);
-			return {
-				prompt,
-				recurring: {
-					mode: "interval",
-					durationMs: normalized.durationMs,
-					note: normalized.note,
-				},
-			};
-		}
-	}
-
-	return {
-		prompt: input,
-		recurring: {
-			mode: "interval",
-			durationMs: DEFAULT_LOOP_INTERVAL,
-		},
-	};
-}
-
-export function parseRemindScheduleArgs(args: string): ReminderParseResult | undefined {
-	const input = args.trim();
-	if (!input) {
-		return undefined;
-	}
-
-	const value = input.toLowerCase().startsWith("in ") ? input.slice(3).trim() : input;
-	const parsed = extractLeadingDuration(value);
-	if (!parsed) {
-		return undefined;
-	}
-
-	const normalized = normalizeDuration(parsed.durationMs);
-	return {
-		prompt: parsed.prompt,
-		durationMs: normalized.durationMs,
-		note: normalized.note,
-	};
-}
-
-export function validateSchedulePromptAddInput(input: { kind?: TaskKind; duration?: string; cron?: string }):
-	| { ok: true; plan: SchedulePromptAddPlan }
-	| {
-			ok: false;
-			error:
-				| "missing_duration"
-				| "invalid_duration"
-				| "invalid_cron_for_once"
-				| "conflicting_schedule_inputs"
-				| "invalid_cron";
-	  } {
-	const kind: TaskKind = input.kind ?? "recurring";
-
-	if (kind === "once") {
-		if (input.cron) {
-			return { ok: false, error: "invalid_cron_for_once" };
-		}
-		if (!input.duration) {
-			return { ok: false, error: "missing_duration" };
-		}
-
-		const parsed = parseDuration(input.duration);
-		if (!parsed) {
-			return { ok: false, error: "invalid_duration" };
-		}
-		const normalized = normalizeDuration(parsed);
-		return { ok: true, plan: { kind: "once", durationMs: normalized.durationMs, note: normalized.note } };
-	}
-
-	if (input.cron && input.duration) {
-		return { ok: false, error: "conflicting_schedule_inputs" };
-	}
-
-	if (input.cron) {
-		const normalizedCron = normalizeCronExpression(input.cron);
-		if (!normalizedCron) {
-			return { ok: false, error: "invalid_cron" };
-		}
-		return {
-			ok: true,
-			plan: {
-				kind: "recurring",
-				mode: "cron",
-				cronExpression: normalizedCron.expression,
-				note: normalizedCron.note,
-			},
-		};
-	}
-
-	if (input.duration) {
-		const parsed = parseDuration(input.duration);
-		if (!parsed) {
-			return { ok: false, error: "invalid_duration" };
-		}
-		const normalized = normalizeDuration(parsed);
-		return {
-			ok: true,
-			plan: { kind: "recurring", mode: "interval", durationMs: normalized.durationMs, note: normalized.note },
-		};
-	}
-
-	return {
-		ok: true,
-		plan: { kind: "recurring", mode: "interval", durationMs: DEFAULT_LOOP_INTERVAL },
-	};
 }
 
 // ── Runtime ─────────────────────────────────────────────────────────────────
@@ -1236,442 +865,6 @@ export class SchedulerRuntime {
 		}
 	}
 }
-
-// ── Commands ────────────────────────────────────────────────────────────────
-
-function registerCommands(pi: ExtensionAPI, runtime: SchedulerRuntime) {
-	pi.registerCommand("loop", {
-		description: "Schedule recurring prompt: /loop 5m <prompt>, /loop <prompt> every 2h, or /loop cron <expr> <prompt>",
-		handler: async (args, ctx) => {
-			const parsed = parseLoopScheduleArgs(args);
-			if (!parsed) {
-				ctx.ui.notify(
-					"Usage: /loop 5m check build OR /loop cron '*/5 * * * *' check build (minimum cron cadence is 1m)",
-					"warning",
-				);
-				return;
-			}
-
-			if (runtime.taskCount >= MAX_TASKS) {
-				ctx.ui.notify(`Task limit reached (${MAX_TASKS}). Delete one with /schedule delete <id>.`, "error");
-				return;
-			}
-
-			if (parsed.recurring.mode === "cron") {
-				const task = runtime.addRecurringCronTask(parsed.prompt, parsed.recurring.cronExpression);
-				if (!task) {
-					ctx.ui.notify("Invalid cron schedule. Cron tasks must run no more often than once per minute.", "error");
-					return;
-				}
-				ctx.ui.notify(`Scheduled cron ${task.cronExpression} (id: ${task.id}). Expires in 3 days.`, "info");
-				if (parsed.recurring.note) {
-					ctx.ui.notify(parsed.recurring.note, "info");
-				}
-				return;
-			}
-
-			const task = runtime.addRecurringIntervalTask(parsed.prompt, parsed.recurring.durationMs);
-			ctx.ui.notify(
-				`Scheduled every ${formatDurationShort(parsed.recurring.durationMs)} (id: ${task.id}). Expires in 3 days.`,
-				"info",
-			);
-			if (parsed.recurring.note) {
-				ctx.ui.notify(parsed.recurring.note, "info");
-			}
-		},
-	});
-
-	pi.registerCommand("remind", {
-		description: "Schedule one-time reminder: /remind in 45m <prompt>",
-		handler: async (args, ctx) => {
-			const parsed = parseRemindScheduleArgs(args);
-			if (!parsed) {
-				ctx.ui.notify("Usage: /remind in 45m check deployment", "warning");
-				return;
-			}
-
-			if (runtime.taskCount >= MAX_TASKS) {
-				ctx.ui.notify(`Task limit reached (${MAX_TASKS}). Delete one with /schedule delete <id>.`, "error");
-				return;
-			}
-
-			const task = runtime.addOneShotTask(parsed.prompt, parsed.durationMs);
-			ctx.ui.notify(`Reminder set for ${runtime.formatRelativeTime(task.nextRunAt)} (id: ${task.id}).`, "info");
-			if (parsed.note) {
-				ctx.ui.notify(parsed.note, "info");
-			}
-		},
-	});
-
-	pi.registerCommand("schedule", {
-		description:
-			"Manage scheduled reminders and future check-ins. No args opens TUI manager. Also: list | enable <id> | disable <id> | delete <id> | clear",
-		// biome-ignore lint/complexity/noExcessiveCognitiveComplexity: Command router with multiple subcommands.
-		handler: async (args, ctx) => {
-			const trimmed = args.trim();
-			if (!trimmed || trimmed === "tui") {
-				await runtime.openTaskManager(ctx);
-				return;
-			}
-
-			const [rawAction, rawArg] = trimmed.split(/\s+/, 2);
-			const action = rawAction.toLowerCase();
-
-			if (action === "list") {
-				pi.sendMessage({
-					customType: "pi-scheduler",
-					content: runtime.formatTaskList(),
-					display: true,
-				});
-				return;
-			}
-
-			if (action === "enable" || action === "disable") {
-				if (!rawArg) {
-					ctx.ui.notify(`Usage: /schedule ${action} <id>`, "warning");
-					return;
-				}
-				const enabled = action === "enable";
-				const ok = runtime.setTaskEnabled(rawArg, enabled);
-				if (!ok) {
-					ctx.ui.notify(`Task not found: ${rawArg}`, "warning");
-					return;
-				}
-				ctx.ui.notify(`${enabled ? "Enabled" : "Disabled"} scheduled task ${rawArg}.`, "info");
-				return;
-			}
-
-			if (action === "delete" || action === "remove" || action === "rm") {
-				if (!rawArg) {
-					ctx.ui.notify("Usage: /schedule delete <id>", "warning");
-					return;
-				}
-				const removed = runtime.deleteTask(rawArg);
-				if (!removed) {
-					ctx.ui.notify(`Task not found: ${rawArg}`, "warning");
-					return;
-				}
-				ctx.ui.notify(`Deleted scheduled task ${rawArg}.`, "info");
-				return;
-			}
-
-			if (action === "clear") {
-				const count = runtime.clearTasks();
-				ctx.ui.notify(`Cleared ${count} task${count === 1 ? "" : "s"}.`, "info");
-				return;
-			}
-
-			ctx.ui.notify("Usage: /schedule [tui|list|enable <id>|disable <id>|delete <id>|clear]", "warning");
-		},
-	});
-
-	pi.registerCommand("unschedule", {
-		description: "Alias for /schedule delete <id>",
-		handler: async (args, ctx) => {
-			const id = args.trim();
-			if (!id) {
-				ctx.ui.notify("Usage: /unschedule <id>", "warning");
-				return;
-			}
-			const removed = runtime.deleteTask(id);
-			if (!removed) {
-				ctx.ui.notify(`Task not found: ${id}`, "warning");
-				return;
-			}
-			ctx.ui.notify(`Deleted scheduled task ${id}.`, "info");
-		},
-	});
-}
-
-// ── Tool ────────────────────────────────────────────────────────────────────
-
-const SchedulePromptToolParams = Type.Object({
-	action: Type.Union(
-		[
-			Type.Literal("add"),
-			Type.Literal("list"),
-			Type.Literal("delete"),
-			Type.Literal("clear"),
-			Type.Literal("enable"),
-			Type.Literal("disable"),
-		],
-		{ description: "Action to perform" },
-	),
-	kind: Type.Optional(Type.Union([Type.Literal("recurring"), Type.Literal("once")], { description: "Task kind" })),
-	prompt: Type.Optional(Type.String({ description: "Prompt text to run when the task fires" })),
-	duration: Type.Optional(
-		Type.String({
-			description:
-				"Delay/interval like 5m, 2h, 1 day. For kind=once this is required. For kind=recurring this creates interval-based loops.",
-		}),
-	),
-	cron: Type.Optional(
-		Type.String({
-			description:
-				"Cron expression for recurring tasks. Accepts 5-field (minute hour dom month dow) or 6-field (sec minute hour dom month dow).",
-		}),
-	),
-	id: Type.Optional(Type.String({ description: "Task id for delete/enable/disable action" })),
-});
-
-function registerTools(pi: ExtensionAPI, runtime: SchedulerRuntime) {
-	pi.registerTool({
-		name: "schedule_prompt",
-		label: "Schedule Prompt",
-		description:
-			"Create/list/enable/disable/delete scheduled prompts. Use this when the user asks for reminders, to check back later, or to follow up on PRs, CI, builds, deployments, or any recurring check. add requires prompt; once tasks require duration; recurring supports interval (duration) or cron expression (cron).",
-		promptSnippet:
-			"Create/list/enable/disable/delete scheduled prompts for one-time reminders, future follow-ups, and recurring PR/CI/build/deployment checks. Supports intervals/cron and one-time reminders while this pi session remains active.",
-		promptGuidelines: [
-			"Use this tool when the user asks to remind/check back later, revisit something in the future, or monitor PRs, CI, builds, deploys, or background work.",
-			"For recurring tasks use kind='recurring' with duration like 5m or 2h, or provide cron.",
-			"For one-time reminders use kind='once' with duration like 30m or 1h.",
-			"Scheduled tasks run only while pi is active and idle, so phrase reminders and follow-ups accordingly.",
-		],
-		parameters: SchedulePromptToolParams,
-		execute: async (
-			_toolCallId,
-			params: { action: string; kind?: TaskKind; prompt?: string; duration?: string; cron?: string; id?: string },
-		): Promise<{ content: { type: "text"; text: string }[]; details: Record<string, unknown> }> => {
-			const { action } = params;
-
-			if (action === "list") {
-				return handleToolList(runtime);
-			}
-			if (action === "clear") {
-				return handleToolClear(runtime);
-			}
-			if (action === "delete") {
-				return handleToolDelete(params, runtime);
-			}
-			if (action === "enable" || action === "disable") {
-				return handleToolToggle(params, runtime);
-			}
-			if (action === "add") {
-				return handleToolAdd(params, runtime);
-			}
-
-			return {
-				content: [{ type: "text", text: `Error: unsupported action '${String(action)}'.` }],
-				details: { action, error: "unsupported_action" },
-			};
-		},
-	});
-}
-
-type ToolResult = { content: { type: "text"; text: string }[]; details: Record<string, unknown> };
-
-function handleToolList(runtime: SchedulerRuntime): ToolResult {
-	const list = runtime.getSortedTasks();
-	if (list.length === 0) {
-		return { content: [{ type: "text", text: "No scheduled tasks." }], details: { action: "list", tasks: [] } };
-	}
-
-	const lines = list.map((task) => {
-		const schedule =
-			task.kind === "once"
-				? "-"
-				: (task.cronExpression ?? formatDurationShort(task.intervalMs ?? DEFAULT_LOOP_INTERVAL));
-		const state = task.enabled ? "on" : "off";
-		const status = task.lastStatus ?? "pending";
-		const last = task.lastRunAt ? runtime.formatRelativeTime(task.lastRunAt) : "never";
-		return `${task.id}\t${state}\t${task.kind}\t${schedule}\t${runtime.formatRelativeTime(task.nextRunAt)}\t${task.runCount}\t${last}\t${status}\t${task.prompt}`;
-	});
-	return {
-		content: [
-			{
-				type: "text",
-				text: `Scheduled tasks (id\tstate\tkind\tschedule\tnext\truns\tlast\tstatus\tprompt):\n${lines.join("\n")}`,
-			},
-		],
-		details: { action: "list", tasks: list },
-	};
-}
-
-function handleToolClear(runtime: SchedulerRuntime): ToolResult {
-	const count = runtime.clearTasks();
-	return {
-		content: [{ type: "text", text: `Cleared ${count} scheduled task${count === 1 ? "" : "s"}.` }],
-		details: { action: "clear", cleared: count },
-	};
-}
-
-function handleToolDelete(params: { id?: string }, runtime: SchedulerRuntime): ToolResult {
-	const id = params.id?.trim();
-	if (!id) {
-		return {
-			content: [{ type: "text", text: "Error: id is required for delete action." }],
-			details: { action: "delete", error: "missing_id" },
-		};
-	}
-	const removed = runtime.deleteTask(id);
-	if (!removed) {
-		return {
-			content: [{ type: "text", text: `Task not found: ${id}` }],
-			details: { action: "delete", id, removed: false },
-		};
-	}
-	return {
-		content: [{ type: "text", text: `Deleted scheduled task ${id}.` }],
-		details: { action: "delete", id, removed: true },
-	};
-}
-
-function handleToolToggle(params: { action: string; id?: string }, runtime: SchedulerRuntime): ToolResult {
-	const { action } = params;
-	const id = params.id?.trim();
-	if (!id) {
-		return {
-			content: [{ type: "text", text: `Error: id is required for ${action} action.` }],
-			details: { action, error: "missing_id" },
-		};
-	}
-	const enabled = action === "enable";
-	const ok = runtime.setTaskEnabled(id, enabled);
-	if (!ok) {
-		return {
-			content: [{ type: "text", text: `Task not found: ${id}` }],
-			details: { action, id, updated: false },
-		};
-	}
-	return {
-		content: [{ type: "text", text: `${enabled ? "Enabled" : "Disabled"} scheduled task ${id}.` }],
-		details: { action, id, updated: true, enabled },
-	};
-}
-
-function validationErrorMessage(error: string): string {
-	switch (error) {
-		case "missing_duration":
-			return "Error: duration is required for one-time reminders.";
-		case "invalid_duration":
-			return "Error: invalid duration. Use values like 5m, 2h, 1 day.";
-		case "invalid_cron_for_once":
-			return "Error: cron is only valid for recurring tasks.";
-		case "conflicting_schedule_inputs":
-			return "Error: provide either duration or cron for recurring tasks, not both.";
-		case "invalid_cron":
-			return "Error: invalid cron expression (minimum cadence is 1 minute).";
-		default:
-			return `Error: ${error}`;
-	}
-}
-
-function handleToolAdd(
-	params: { kind?: TaskKind; prompt?: string; duration?: string; cron?: string },
-	runtime: SchedulerRuntime,
-): ToolResult {
-	const prompt = params.prompt?.trim();
-	if (!prompt) {
-		return {
-			content: [{ type: "text", text: "Error: prompt is required for add action." }],
-			details: { action: "add", error: "missing_prompt" },
-		};
-	}
-
-	if (runtime.taskCount >= MAX_TASKS) {
-		return {
-			content: [{ type: "text", text: `Task limit reached (${MAX_TASKS}). Delete one first.` }],
-			details: { action: "add", error: "task_limit" },
-		};
-	}
-
-	const validated = validateSchedulePromptAddInput({
-		kind: params.kind,
-		duration: params.duration,
-		cron: params.cron,
-	});
-	if (!validated.ok) {
-		return {
-			content: [{ type: "text", text: validationErrorMessage(validated.error) }],
-			details: { action: "add", error: validated.error },
-		};
-	}
-
-	if (validated.plan.kind === "once") {
-		const task = runtime.addOneShotTask(prompt, validated.plan.durationMs);
-		return {
-			content: [
-				{
-					type: "text",
-					text: `Reminder scheduled (id: ${task.id}) for ${runtime.formatRelativeTime(task.nextRunAt)}.${
-						validated.plan.note ? ` ${validated.plan.note}` : ""
-					}`,
-				},
-			],
-			details: { action: "add", task },
-		};
-	}
-
-	if (validated.plan.mode === "cron") {
-		const task = runtime.addRecurringCronTask(prompt, validated.plan.cronExpression);
-		if (!task) {
-			return {
-				content: [
-					{ type: "text", text: "Error: invalid cron expression or cadence is too frequent (minimum is 1 minute)." },
-				],
-				details: { action: "add", error: "cron_next_run_failed" },
-			};
-		}
-		return {
-			content: [
-				{
-					type: "text",
-					text: `Recurring cron task scheduled (id: ${task.id}) with '${task.cronExpression}'. Expires in 3 days.${
-						validated.plan.note ? ` ${validated.plan.note}` : ""
-					}`,
-				},
-			],
-			details: { action: "add", task },
-		};
-	}
-
-	const task = runtime.addRecurringIntervalTask(prompt, validated.plan.durationMs);
-	return {
-		content: [
-			{
-				type: "text",
-				text: `Recurring task scheduled (id: ${task.id}) every ${formatDurationShort(validated.plan.durationMs)}. Expires in 3 days.${
-					validated.plan.note ? ` ${validated.plan.note}` : ""
-				}`,
-			},
-		],
-		details: { action: "add", task },
-	};
-}
-
-// ── Events ──────────────────────────────────────────────────────────────────
-
-function registerEvents(pi: ExtensionAPI, runtime: SchedulerRuntime) {
-	pi.on("session_start", async (_event, ctx) => {
-		runtime.setRuntimeContext(ctx);
-		runtime.startScheduler();
-		runtime.updateStatus();
-	});
-
-	pi.on("session_switch", async (_event, ctx) => {
-		runtime.setRuntimeContext(ctx);
-		runtime.updateStatus();
-	});
-
-	pi.on("session_fork", async (_event, ctx) => {
-		runtime.setRuntimeContext(ctx);
-		runtime.updateStatus();
-	});
-
-	pi.on("session_tree", async (_event, ctx) => {
-		runtime.setRuntimeContext(ctx);
-		runtime.updateStatus();
-	});
-
-	pi.on("session_shutdown", async (_event, ctx) => {
-		runtime.setRuntimeContext(ctx);
-		runtime.stopScheduler();
-		runtime.clearStatus(ctx);
-	});
-}
-
-// ── Extension entry ─────────────────────────────────────────────────────────
 
 export default function schedulerExtension(pi: ExtensionAPI) {
 	const runtime = new SchedulerRuntime(pi);
